Break up atmosphere class and composition into separate files.
