+++ DataDefinitions/AtmosphereClass.cs
@@ -6,67 +6,7 @@ using System.Resources;
 
 namespace EddiDataDefinitions
 {
-    /// <summary> Atmosphere Composition </summary>
-    [JsonObject(MemberSerialization.OptIn)]
-    public class AtmosphereComposition
-    {
-        static AtmosphereComposition()
-        {
-            resourceManager = Properties.AtmosphereComposition.ResourceManager;
-            resourceManager.IgnoreCase = true;
-
-            COMPOSITIONS.Add("water", "Water");
-            COMPOSITIONS.Add("oxygen", "Oxygen");
-            COMPOSITIONS.Add("carbondioxide", "Carbon dioxide");
-            COMPOSITIONS.Add("sulphurdioxide", "Sulphur dioxide");
-            COMPOSITIONS.Add("ammonia", "Ammonia");
-            COMPOSITIONS.Add("methane", "Methane");
-            COMPOSITIONS.Add("nitrogen", "Nitrogen");
-            COMPOSITIONS.Add("hydrogen", "Hydrogen");
-            COMPOSITIONS.Add("helium", "Helium");
-            COMPOSITIONS.Add("neon", "Neon");
-            COMPOSITIONS.Add("argon", "Argon");
-            COMPOSITIONS.Add("silicates", "Silicates");
-            COMPOSITIONS.Add("iron", "Iron");
-        }
-
-        public static readonly ResourceManager resourceManager;
-
-        // Translation of composition of atmosphere 
-        private static readonly IDictionary<string, string> COMPOSITIONS = new Dictionary<string, string>();
-
-        [JsonProperty("composition")]
-        public string edName { get; set; } // Nitrogen, Oxygen, etc.
-        public string invariantName => GetInvariantString(edName);
-        public string localizedName => GetLocalizedString(edName);
-        [JsonIgnore, Obsolete("Please use localizedName or invariantName")]
-        public string name => localizedName;
-
-        [JsonProperty("share")]
-        public decimal percent { get; set; } // Percent share of the atmosphere
-
-        private string GetInvariantString(string name)
-        {
-            if (name == null) { return null; }
-            name = name.Replace(" ", "_");
-            return resourceManager.GetString(name, CultureInfo.InvariantCulture);
-        }
-
-        private string GetLocalizedString(string name)
-        {
-            if (name == null) { return null; }
-            name = name.Replace(" ", "_");
-            return resourceManager.GetString(name);
-        }
-
-        public AtmosphereComposition(string composition, decimal percent)
-        {
-            this.edName = composition;
-            this.percent = percent;
-        }
-    }
-
-    /// <summary> Atmosphere Composition </summary>
+    /// <summary> Atmosphere Class </summary>
     [JsonObject(MemberSerialization.OptIn)]
     public class AtmosphereClass : ResourceBasedLocalizedEDName<AtmosphereClass>
     {
@@ -76,7 +16,7 @@ namespace EddiDataDefinitions
             resourceManager.IgnoreCase = true;
             missingEDNameHandler = (edname) => new AtmosphereClass(edname);
 
-            None = new AtmosphereClass("None");
+            var None = new AtmosphereClass("None");
             var Ammonia = new AtmosphereClass("Ammonia");
             var AmmoniaAndOxygen = new AtmosphereClass("AmmoniaAndOxygen");
             var AmmoniaRich = new AtmosphereClass("AmmoniaRich");
@@ -100,10 +40,8 @@ namespace EddiDataDefinitions
             var WaterRich = new AtmosphereClass("WaterRich");
         }
 
-        public static readonly AtmosphereClass None;
-
         // dummy used to ensure that the static constructor has run, defaulting to "None" if no atmosphere is specified.
-        public AtmosphereClass() : this("None")
+        public AtmosphereClass() : this("")
         { }
 
         private AtmosphereClass(string edname) : base(edname, edname
@@ -119,7 +57,7 @@ namespace EddiDataDefinitions
         {
             if (name == null)
             {
-                return None;
+                return FromName("None");
             }
 
             // Temperature and pressure are defined separately so we remove them from this string (if descriptors are present)
@@ -135,7 +73,7 @@ namespace EddiDataDefinitions
         {
             if (edname == null)
             {
-                return None;
+                return FromEDName("None");
             }
 
             // Temperature and pressure are defined separately so we remove them from this string (if descriptors are present)
